# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mr Analyst API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TipTier = typeof TipTier[keyof typeof TipTier];


export const TipTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export type TipStatus = typeof TipStatus[keyof typeof TipStatus];


export const TipStatus = {
  locked: 'locked',
  pending: 'pending',
  won: 'won',
  lost: 'lost',
  postponed: 'postponed',
  cancelled: 'cancelled',
} as const;

export interface Tip {
  id: string;
  tier: TipTier;
  /** @nullable */
  teams?: string | null;
  /** @nullable */
  tip_type?: string | null;
  /** @nullable */
  odds?: number | null;
  status: TipStatus;
  /** @nullable */
  match_date?: string | null;
  /** @nullable */
  match_time?: string | null;
  /** @nullable */
  publish_at?: string | null;
  is_locked: boolean;
  created_at: string;
}

export type TipInputTier = typeof TipInputTier[keyof typeof TipInputTier];


export const TipInputTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export type TipInputStatus = typeof TipInputStatus[keyof typeof TipInputStatus];


export const TipInputStatus = {
  locked: 'locked',
  pending: 'pending',
  won: 'won',
  lost: 'lost',
  postponed: 'postponed',
  cancelled: 'cancelled',
} as const;

export interface TipInput {
  tier: TipInputTier;
  teams?: string;
  tip_type?: string;
  odds?: number;
  status: TipInputStatus;
  match_date?: string;
  match_time?: string;
  publish_at?: string;
}

export type TipUpdateTier = typeof TipUpdateTier[keyof typeof TipUpdateTier];


export const TipUpdateTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export type TipUpdateStatus = typeof TipUpdateStatus[keyof typeof TipUpdateStatus];


export const TipUpdateStatus = {
  locked: 'locked',
  pending: 'pending',
  won: 'won',
  lost: 'lost',
  postponed: 'postponed',
  cancelled: 'cancelled',
} as const;

export interface TipUpdate {
  tier?: TipUpdateTier;
  teams?: string;
  tip_type?: string;
  odds?: number;
  status?: TipUpdateStatus;
  match_date?: string;
  match_time?: string;
  publish_at?: string;
}

export interface TierStats {
  total: number;
  won: number;
  lost: number;
  pending: number;
  postponed: number;
  cancelled: number;
  win_rate: number;
  todays_tips: number;
}

export interface Stats {
  pro_plus: TierStats;
  pro: TierStats;
}

export interface User {
  id: string;
  email: string;
  /** @nullable */
  subscription?: string | null;
  created_at: string;
  /** @nullable */
  last_sign_in_at?: string | null;
  is_admin: boolean;
}

export interface UserUpdate {
  /** @nullable */
  subscription?: string | null;
  is_admin?: boolean;
}

export interface AppConfig {
  key: string;
  value: string;
  /** @nullable */
  description?: string | null;
  updated_at: string;
}

export interface AppConfigUpdate {
  key: string;
  value: string;
}

export type LiveMessageTier = typeof LiveMessageTier[keyof typeof LiveMessageTier];


export const LiveMessageTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export interface LiveMessage {
  id: string;
  user_name: string;
  /** @nullable */
  user_city?: string | null;
  message: string;
  tier: LiveMessageTier;
  created_at: string;
}

export type LiveMessageInputTier = typeof LiveMessageInputTier[keyof typeof LiveMessageInputTier];


export const LiveMessageInputTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export interface LiveMessageInput {
  user_name: string;
  user_city?: string;
  message: string;
  tier: LiveMessageInputTier;
}

export type VipLoginInputTier = typeof VipLoginInputTier[keyof typeof VipLoginInputTier];


export const VipLoginInputTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export interface VipLoginInput {
  email: string;
  access_code: string;
  tier?: VipLoginInputTier;
}

export interface VipLoginResult {
  success: boolean;
  /** @nullable */
  tier?: string | null;
  message: string;
  /** @nullable */
  token?: string | null;
}

export type ScheduledPostTier = typeof ScheduledPostTier[keyof typeof ScheduledPostTier];


export const ScheduledPostTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export type ScheduledPostStatus = typeof ScheduledPostStatus[keyof typeof ScheduledPostStatus];


export const ScheduledPostStatus = {
  scheduled: 'scheduled',
  published: 'published',
  cancelled: 'cancelled',
} as const;

export interface ScheduledPost {
  id: string;
  title: string;
  content: string;
  tier: ScheduledPostTier;
  publish_at: string;
  status: ScheduledPostStatus;
  created_at: string;
}

export type ScheduledPostInputTier = typeof ScheduledPostInputTier[keyof typeof ScheduledPostInputTier];


export const ScheduledPostInputTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export interface ScheduledPostInput {
  title: string;
  content: string;
  tier: ScheduledPostInputTier;
  publish_at: string;
}

export type ScheduledPostUpdateTier = typeof ScheduledPostUpdateTier[keyof typeof ScheduledPostUpdateTier];


export const ScheduledPostUpdateTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export type ScheduledPostUpdateStatus = typeof ScheduledPostUpdateStatus[keyof typeof ScheduledPostUpdateStatus];


export const ScheduledPostUpdateStatus = {
  scheduled: 'scheduled',
  published: 'published',
  cancelled: 'cancelled',
} as const;

export interface ScheduledPostUpdate {
  title?: string;
  content?: string;
  tier?: ScheduledPostUpdateTier;
  publish_at?: string;
  status?: ScheduledPostUpdateStatus;
}

export type GetTipsParams = {
tier?: GetTipsTier;
limit?: number;
offset?: number;
};

export type GetTipsTier = typeof GetTipsTier[keyof typeof GetTipsTier];


export const GetTipsTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;

export type GetLiveMessagesParams = {
limit?: number;
};

export type AdminGetTipsParams = {
tier?: AdminGetTipsTier;
limit?: number;
offset?: number;
};

export type AdminGetTipsTier = typeof AdminGetTipsTier[keyof typeof AdminGetTipsTier];


export const AdminGetTipsTier = {
  pro_plus: 'pro_plus',
  pro: 'pro',
} as const;