# lib/api-zod/src/generated/types/scheduledPostInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mr Analyst API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ScheduledPostInputTier } from './scheduledPostInputTier';

export interface ScheduledPostInput {
  title: string;
  content: string;
  tier: ScheduledPostInputTier;
  publish_at: string;
}