# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Mr Analyst API specification
 * OpenAPI spec version: 0.1.0
 */

export interface User {
  id: string;
  email: string;
  /** @nullable */
  subscription?: string | null;
  created_at: string;
  /** @nullable */
  last_sign_in_at?: string | null;
  is_admin: boolean;
}